+++ src/main/java/org/markdownwriterfx/options/GeneralOptionsPane.java
@@ -34,6 +34,7 @@ import java.util.SortedMap;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
+import org.markdownwriterfx.Messages;
 import org.markdownwriterfx.util.Item;
 import org.tbee.javafx.scene.layout.fxml.MigPane;
 
@@ -52,9 +53,9 @@ public class GeneralOptionsPane
 		String defaultLineSeparator = System.getProperty( "line.separator", "\n" );
 		String defaultLineSeparatorStr = defaultLineSeparator.replace("\r", "CR").replace("\n", "LF");
 		lineSeparatorField.getItems().addAll(
-			new Item<String>( "Platform Default (" + defaultLineSeparatorStr + ')', null ),
-			new Item<String>( "Windows (CRLF)", "\r\n" ),
-			new Item<String>( "Unix (LF)", "\n" ));
+			new Item<String>( Messages.get("GeneralOptionsPane.platformDefault", defaultLineSeparatorStr), null ),
+			new Item<String>( Messages.get("GeneralOptionsPane.sepWindows"), "\r\n" ),
+			new Item<String>( Messages.get("GeneralOptionsPane.sepUnix"), "\n" ));
 
 		encodingField.getItems().addAll(getAvailableEncodings());
 	}
@@ -63,7 +64,7 @@ public class GeneralOptionsPane
 		SortedMap<String, Charset> availableCharsets = Charset.availableCharsets();
 
 		ArrayList<Item<String>> values = new ArrayList<>(1 + availableCharsets.size());
-		values.add(new Item<String>("Platform Default (" + Charset.defaultCharset().name() + ')', null));
+		values.add(new Item<String>(Messages.get("GeneralOptionsPane.platformDefault", Charset.defaultCharset().name()), null));
 		for (String name : availableCharsets.keySet())
 			values.add(new Item<String>(name, name));
 		return values;
@@ -100,17 +101,17 @@ public class GeneralOptionsPane
 		setRows("[][]para[][]");
 
 		//---- lineSeparatorLabel ----
-		lineSeparatorLabel.setText("_Line separator:");
+		lineSeparatorLabel.setText(Messages.get("GeneralOptionsPane.lineSeparatorLabel.text"));
 		lineSeparatorLabel.setMnemonicParsing(true);
 		add(lineSeparatorLabel, "cell 0 0");
 		add(lineSeparatorField, "cell 1 0");
 
 		//---- lineSeparatorLabel2 ----
-		lineSeparatorLabel2.setText("(applies to new files only)");
+		lineSeparatorLabel2.setText(Messages.get("GeneralOptionsPane.lineSeparatorLabel2.text"));
 		add(lineSeparatorLabel2, "cell 2 0");
 
 		//---- encodingLabel ----
-		encodingLabel.setText("En_coding:");
+		encodingLabel.setText(Messages.get("GeneralOptionsPane.encodingLabel.text"));
 		encodingLabel.setMnemonicParsing(true);
 		add(encodingLabel, "cell 0 1");
 
@@ -119,7 +120,7 @@ public class GeneralOptionsPane
 		add(encodingField, "cell 1 1");
 
 		//---- showWhitespaceCheckBox ----
-		showWhitespaceCheckBox.setText("Show _Whitespace Characters");
+		showWhitespaceCheckBox.setText(Messages.get("GeneralOptionsPane.showWhitespaceCheckBox.text"));
 		add(showWhitespaceCheckBox, "cell 0 2 3 1,growx 0,alignx left");
 
 		//---- spellCheckerCheckBox ----