+++ CHANGES.md
@@ -1,5 +1,11 @@
 Markdown Writer FX Change Log
 =============================
 
+## 0.2
+- RichTextFX (and dependencies) updated to versoin 0.6.10 (fixes bugs and memory leaks).
+- pegdown Markdown parser updated to version 1.6.
+- Added five new pegdown 1.6 extension flags to Markdown Options tab.
+- Minor improvements.
+
 ## 0.1
 - Initial release